+++ core/CHANGELOG.md
@@ -20,15 +20,18 @@
 ### 🤖 PixiJS Engine
 
 - 🆕 [v1.0.0] `components/componentGenerator`.
-- 🆕 [v1.0.0] `components/actorGenerator`.
+- 🆕 [v1.0.0] `components/blockGenerator`.
 - 🆕 [v1.0.0] `components/functionalities/hitTheWallFunctionalityBuilder`.
 - 🆕 [v1.0.0] `components/functionalities/positioningFunctionalityBuilder`.
+- 🆕 [v1.0.0] `components/functionalities/emotionFunctionalityBuilder`.
+- 🆕 [v1.0.0] `components/functionalities/theFallenOneFunctionalityBuilder`.
 - 🆕 [v1.0.0] `engineGenerator`.
 - 🆕 [v1.0.0] `engineGenerator/processors/processorGenerator`.
 - 🆕 [v1.0.0] `engineGenerator/processors/positioningProcessorBuilder`.
 - 🆕 [v1.0.0] `engineGenerator/processors/checkers/engineErrorBuilder`.
 - 🆕 [v1.0.0] `engineGenerator/processors/checkers/hasHitAWallBuilder`.
 - 🆕 [v1.0.0] `engineGenerator/processors/checkers/hasHitAWallErrorGenerator`.
+- 🆕 [v1.0.0] `generics/actorGenerator`.
 
 ### 💅 React UI